refactor: clean app shell map offsets
